Rename Post components
